Changes:
- remove All comments by perant id concatMap, update comment-component.
- get All accounts, update accounts-services and add-edit-bug-component.
- update add-edit-comment-component.

+++ src/app/_bugs/components/bug/bug.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { Router } from '@angular/router';
 import { BugsService } from 'src/app/core/data-access/bugs.service';
 import { Bug } from '../../data-access/bug';
+import { CommentsService } from 'src/app/core/data-access/comments.service';
 
 @Component({
   selector: 'app-bug',
@@ -12,7 +13,11 @@ export class BugComponent implements OnInit {
   @Input() bug!: any;
   @Output() removeBugEvent = new EventEmitter<Bug>();
 
-  constructor(private bugsService: BugsService, private router: Router) {}
+  constructor(
+    private bugsService: BugsService,
+    private commentsService: CommentsService,
+    private router: Router
+  ) {}
 
   ngOnInit(): void {
     console.log('from ngoninit bugcomponent', this.bug);
@@ -23,10 +28,27 @@ export class BugComponent implements OnInit {
   }
 
   removeBug() {
-    this.bugsService.removeBug(this.bug.bug_id).subscribe(() => {
-      this.removeBugEvent.emit(this.bug);
-      console.log('from removebug subscribe');
-    });
+    if (
+      confirm(
+        'Are you sure to delete this Bug? you will delete also all it related comments '
+      )
+    ) {
+      this.bugsService.removeBug(this.bug.bug_id).subscribe(() => {
+        this.deleteAllCommentsRelatedToBug();
+        //ToDo Merge two Observables into one Observable of the same type (2 delete requset to one subscribe)
+        //Do multiple HTTP requests using the RxJs Concat Operator
+        //https://www.youtube.com/watch?v=b59tcUwfpWU
+      });
+    }
+  }
+
+  deleteAllCommentsRelatedToBug() {
+    this.commentsService
+      .removeCommentsByBugId(this.bug.bug_id)
+      .subscribe(() => {
+        this.removeBugEvent.emit(this.bug);
+        console.log('something went wrong');
+      });
   }
 
   bugDetails() {
